backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.7
SpeechRecognition==3.10.1
pydub==0.25.1
pyaudio==0.2.14
ffmpeg-python==0.2.0
librosa==0.10.1
soundfile==0.12.1
numpy>=1.20.0
torch>=2.0.0
transformers>=4.30.0
scipy==1.10.1
scikit-learn>=1.0.0
openai-whisper>=20231117
accelerate>=0.20.0
sentencepiece>=0.1.99
tiktoken>=0.4.0
